- Changed the encoding for ALL language texfiles to UTF 8 o Now we can also build FUNCTIONAL Polish version with VS 2005 instead of only building with VS 2008/VS 2010
- Multiplayer fixes and improvements
o Increased MP version to v1.9
o Added and info output text when the AI interrupts the pure client, that the game could hang and the server should press ALT + E
o Disabled the possibility, that the server could always give himself the turn with ALT + E. Now when pressing '1' on the dialog instead of giving the server the turn, the dialog closes without any action
o Disabled the possibility, that the server could kick himself from the game when pressing ALT + K and then selecting '1'. The '1' button now cancels the dialog without doing any action.
o Disabled the possibility, that the server could kick a client in the strategy screen
o When client gets kicked from tactical, the game automatically switches to spectator mode for the kicked client
o Disabled Encyclopedia and Briefing room links from the laptop

git-svn-id: https://ja2svn.dyndns.org/source/ja2/trunk/GameSource/ja2_v1.13@4692 3b4a5df2-a311-0410-b5c6-a8a6f20db521