New modifiers for Item Properties
The new Item_Settings.ini is there to host modifiers for item properties in the different XML files. For now it is only used for the Weapon Overheating feature to adjust parameters based on gun class.
It uses the new struct gItemSettings which is populated in function void LoadItemSettings() on game start.

git-svn-id: https://ja2svn.dyndns.org/source/ja2/trunk/GameSource/ja2_v1.13@6463 3b4a5df2-a311-0410-b5c6-a8a6f20db521